requirements.txt
appdirs==1.4.3
asgiref==3.2.3
black==20.8b1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.0.7
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
inflection==0.3.1
ipaddr==2.2.0
itypes==1.1.0
Jinja2==2.11.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mypy-extensions==0.4.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
pathspec==0.8.1
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2018.9
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
